Probably switch to esbuild
