+++ kolmafia-polyfill.js
@@ -1,0 +1,2 @@
+const kolmafia = require("kolmafia");
+export let console = { log: kolmafia.print };
