+++ package.json
@@ -5,9 +5,7 @@
     "main": "phredhccs.js",
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1",
-        "build": "npm run build:types && npm run build:js",
-        "build:types": "tsc",
-        "build:js": "webpack",
+        "build": "node build.mjs",
         "lint": "eslint src && prettier --check src",
         "watch": "webpack --watch --progress"
     },
@@ -15,7 +13,7 @@
         "@babel/cli": "^7.14.8",
         "@babel/compat-data": "^7.15.0",
         "@babel/core": "^7.15.0",
-        "@babel/plugin-proposal-class-properties": "^7.14.5",
+        "@babel/plugin-proposal-class-properties": "^7.18.6",
         "@babel/plugin-proposal-decorators": "^7.14.5",
         "@babel/plugin-proposal-object-rest-spread": "^7.14.7",
         "@babel/preset-env": "^7.15.0",
@@ -23,13 +21,14 @@
         "@typescript-eslint/eslint-plugin": "^4.29.2",
         "@typescript-eslint/parser": "^4.29.2",
         "babel-loader": "^8.2.2",
+        "esbuild": "0.17.0",
+        "esbuild-plugin-babel": "0.2.3",
         "eslint": "^7.32.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-libram": "^0.2.26",
+        "eslint-plugin-prettier": "4.2.1",
         "prettier": "^2.3.1",
-        "typescript": "^4.0.3",
-        "webpack": "^5.8.0",
-        "webpack-cli": "^4.2.0"
+        "typescript": "^4.5.2"
     },
     "dependencies": {
         "grimoire-kolmafia": "^0.3.11",
