--- patches/@babel+compat-data+7.12.7.patch
@@ -1,344 +1,0 @@
-diff --git a/node_modules/@babel/compat-data/data/corejs2-built-ins.json b/node_modules/@babel/compat-data/data/corejs2-built-ins.json
-index 430cdbf..e817bb9 100644
---- a/node_modules/@babel/compat-data/data/corejs2-built-ins.json
-+++ b/node_modules/@babel/compat-data/data/corejs2-built-ins.json
-@@ -8,6 +8,7 @@
-     "node": "4",
-     "ios": "9",
-     "samsung": "5",
-+    "rhino": "1.7.13",
-     "electron": "0.31"
-   },
-   "es6.array.every": {
-@@ -22,6 +23,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.array.fill": {
-@@ -33,6 +35,7 @@
-     "node": "4",
-     "ios": "8",
-     "samsung": "5",
-+    "rhino": "1.7.13",
-     "electron": "0.31"
-   },
-   "es6.array.filter": {
-@@ -47,6 +50,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.array.find": {
-@@ -58,6 +62,7 @@
-     "node": "4",
-     "ios": "8",
-     "samsung": "5",
-+    "rhino": "1.7.13",
-     "electron": "0.31"
-   },
-   "es6.array.find-index": {
-@@ -69,6 +74,7 @@
-     "node": "4",
-     "ios": "8",
-     "samsung": "5",
-+    "rhino": "1.7.13",
-     "electron": "0.31"
-   },
-   "es7.array.flat-map": {
-@@ -94,6 +100,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.array.from": {
-@@ -130,6 +137,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.array.is-array": {
-@@ -144,6 +152,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.array.iterator": {
-@@ -155,6 +164,7 @@
-     "node": "10",
-     "ios": "9",
-     "samsung": "9",
-+    "rhino": "1.7.13",
-     "electron": "3.0"
-   },
-   "es6.array.last-index-of": {
-@@ -169,6 +179,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.array.map": {
-@@ -183,6 +194,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.array.of": {
-@@ -194,6 +206,7 @@
-     "node": "4",
-     "ios": "9",
-     "samsung": "5",
-+    "rhino": "1.7.13",
-     "electron": "0.31"
-   },
-   "es6.array.reduce": {
-@@ -208,6 +221,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.array.reduce-right": {
-@@ -222,6 +236,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.array.some": {
-@@ -236,6 +251,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.array.sort": {
-@@ -273,6 +289,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.date.to-iso-string": {
-@@ -287,6 +304,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.date.to-json": {
-@@ -300,6 +318,7 @@
-     "android": "4",
-     "ios": "10",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.date.to-primitive": {
-@@ -325,6 +344,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.function.bind": {
-@@ -339,6 +359,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.function.has-instance": {
-@@ -363,6 +384,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.map": {
-@@ -473,6 +495,7 @@
-     "node": "0.12",
-     "ios": "8",
-     "samsung": "3",
-+    "rhino": "1.7.13",
-     "electron": "0.20"
-   },
-   "es6.math.imul": {
-@@ -596,6 +619,7 @@
-     "android": "4.1",
-     "ios": "9",
-     "samsung": "1.5",
-+    "rhino": "1.7.13",
-     "electron": "0.20"
-   },
-   "es6.number.is-integer": {
-@@ -607,6 +631,7 @@
-     "node": "0.12",
-     "ios": "9",
-     "samsung": "2",
-+    "rhino": "1.7.13",
-     "electron": "0.20"
-   },
-   "es6.number.is-nan": {
-@@ -619,6 +644,7 @@
-     "android": "4.1",
-     "ios": "9",
-     "samsung": "1.5",
-+    "rhino": "1.7.13",
-     "electron": "0.20"
-   },
-   "es6.number.is-safe-integer": {
-@@ -630,6 +656,7 @@
-     "node": "0.12",
-     "ios": "9",
-     "samsung": "2",
-+    "rhino": "1.7.13",
-     "electron": "0.20"
-   },
-   "es6.number.max-safe-integer": {
-@@ -641,6 +668,7 @@
-     "node": "0.12",
-     "ios": "9",
-     "samsung": "2",
-+    "rhino": "1.7.13",
-     "electron": "0.20"
-   },
-   "es6.number.min-safe-integer": {
-@@ -652,6 +680,7 @@
-     "node": "0.12",
-     "ios": "9",
-     "samsung": "2",
-+    "rhino": "1.7.13",
-     "electron": "0.20"
-   },
-   "es6.number.parse-float": {
-@@ -699,6 +728,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es7.object.define-getter": {
-@@ -735,6 +765,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es6.object.define-properties": {
-@@ -749,6 +780,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es7.object.entries": {
-@@ -871,6 +903,7 @@
-     "android": "4.1",
-     "ios": "9",
-     "samsung": "1.5",
-+    "rhino": "1.7.13",
-     "electron": "0.20"
-   },
-   "es6.object.is-frozen": {
-@@ -1287,6 +1320,7 @@
-     "node": "4",
-     "ios": "9",
-     "samsung": "3.4",
-+    "rhino": "1.7.13",
-     "electron": "0.21"
-   },
-   "es6.string.ends-with": {
-@@ -1298,6 +1332,7 @@
-     "node": "4",
-     "ios": "9",
-     "samsung": "3.4",
-+    "rhino": "1.7.13",
-     "electron": "0.21"
-   },
-   "es6.string.fixed": {
-@@ -1348,6 +1383,7 @@
-     "node": "4",
-     "ios": "9",
-     "samsung": "3.4",
-+    "rhino": "1.7.13",
-     "electron": "0.21"
-   },
-   "es6.string.includes": {
-@@ -1359,6 +1395,7 @@
-     "node": "4",
-     "ios": "9",
-     "samsung": "3.4",
-+    "rhino": "1.7.13",
-     "electron": "0.21"
-   },
-   "es6.string.italics": {
-@@ -1440,6 +1477,7 @@
-     "node": "4",
-     "ios": "9",
-     "samsung": "3.4",
-+    "rhino": "1.7.13",
-     "electron": "0.21"
-   },
-   "es6.string.small": {
-@@ -1464,6 +1502,7 @@
-     "node": "4",
-     "ios": "9",
-     "samsung": "3.4",
-+    "rhino": "1.7.13",
-     "electron": "0.21"
-   },
-   "es6.string.strike": {
-@@ -1517,6 +1556,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7",
-     "electron": "0.20"
-   },
-   "es7.string.trim-left": {
-@@ -1564,6 +1604,7 @@
-     "ios": "6",
-     "phantom": "2",
-     "samsung": "1",
-+    "rhino": "1.7.13",
-     "electron": "0.20"
-   },
-   "es6.typed.int8-array": {
-diff --git a/node_modules/@babel/compat-data/data/plugin-bugfixes.json b/node_modules/@babel/compat-data/data/plugin-bugfixes.json
-index e7c13dc..6e10fe6 100644
---- a/node_modules/@babel/compat-data/data/plugin-bugfixes.json
-+++ b/node_modules/@babel/compat-data/data/plugin-bugfixes.json
-@@ -98,6 +98,7 @@
-     "ie": "11",
-     "ios": "11",
-     "samsung": "5",
-+    "rhino": "1.7.13",
-     "electron": "0.37"
-   },
-   "transform-template-literals": {
